% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer_presentation_CJK.R
\name{beamer_presentation_CJK}
\alias{beamer_presentation_CJK}
\title{R Markdown 上で XeLaTeX を使い日本語 beamer スライドを作成するフォーマット}
\usage{
beamer_presentation_CJK(
  theme = "default",
  theme_options = "progressbar=frametitle,block=fill",
  fonttheme = "professionalfonts",
  colortheme = "default",
  innertheme = "default",
  outertheme = "default",
  fontsize = "12pt",
  fig_width = 5.03937,
  fig_height = 3.77953,
  out.width = "100\%",
  out.height = "100\%",
  fig_crop = TRUE,
  fig_caption = TRUE,
  linkcolor = "blue",
  citecolor = "blue",
  urlcolor = "magenta",
  highlight = "default",
  citation_package = "natbib",
  citation_options = "numbers",
  figurename = "図",
  tablename = "表",
  incremental = FALSE,
  keep_tex = TRUE,
  keep_md = FALSE,
  self_contained = TRUE
)
}
\arguments{
\item{theme}{chracter. beamer テーマ. デフォルトは "default"}

\item{theme_options}{character. テーマオプション. デフォルトはフレームタイトルの下にプログレスバーをつけて, ブロックの背景色を描画するというもの}

\item{fonttheme}{character. フォントテーマ. デフォルトでは数式にローマン体を使う. デフォルト: "professional"}

\item{colortheme}{character. 色テーマ. デフォルト: "default"}

\item{innertheme}{character.  \code{colortheme} に同じ}

\item{outertheme}{character. \code{outertheme} に同じ}

\item{fontsize}{character. beamer の規定フォントサイズ. デフォルト: "11pt"}

\item{fig_width}{numeric. 画像保存時の幅. 単位インチ. デフォルトはbeamerのデフォルト幅と同じ. デフォルト: 5.03937}

\item{fig_height}{numeric. \code{fig_width} 参照. デフォルト: 3.77953}

\item{out.width}{character. 画像を貼り付ける際のサイズ. チャンクごとに指定することも可能. デフォルト: "100\%"}

\item{fig_crop}{logical. pdfcrop を使ってpdf画像の余白を削るかどうか. デフォルト: TRUE}

\item{fig_caption}{logical. 画像にキャプションを付けるか否か. デフォルト: TRUE}

\item{linkcolor}{character. 相互参照リンクの色. デフォルト: blue}

\item{citecolor}{character. 参考文献リストへのリンクの色. デフォルト: blue}

\item{urlcolor}{character. URLリンクの色. デフォルト: magenta}

\item{highlight}{character. シンタックスハイライトのデザイン. \code{default}, \code{tango}, \code{pygments}, \code{kate}, \code{monochrome}, \code{espresso}, \code{zenburn}, \code{haddock}, \code{breezedark}, \code{textmate} から指定する.}

\item{citation_package}{character.  本文中の引用トークンに関するパッケージ. デフォルト: natbib}

\item{citation_options}{character. \code{citation_package} のオプション. デフォルトの natbib+numberでは "\link{1}" のような引用トークンが生成される. デフォルト: numbers.}

\item{figurename}{character. 図X の「図」の部分のテキスト. デフォルト: "図"}

\item{tablename}{character. 表Y の「表」の部分のテキスト. デフォルト: "表"}

\item{incremental}{logical. 箇条書きが順番に現れるやつ. 文字が回転するアニメーション機能はない. デフォルト: FALSE『...遠慮のないマッポ関係者が失笑した。ナムサン！プレゼンテーションにおける典型的なセンスレス文字操作だ。』--- B. ボンド& F. モーゼズ}

\item{keep_tex}{logical. 出力時に .tex ファイルを残すかどうか. デバッグのため TRUEを推奨する. デフォルト: TRUE}

\item{keep_md}{logical. 出力時に .md ファイルを残すかどうか. デフォルト: FALSE}

\item{self_contained}{Whether to generate a full LaTeX document (\code{TRUE})
or just the body of a LaTeX document (\code{FALSE}). Note the LaTeX
document is an intermediate file unless \code{keep_tex = TRUE}.}

\item{out.heigt}{character. \code{out.height} 参照. デフォルト: "100\%"}
}
\value{
rmarkdown_output_format
}
\description{
rmarkdownで \LaTeX を使う場合, 日本語を適切に表示するためにいろいろ必要だった調整を済ませたフォーマット
基本的なオプションだけに限定することで簡単にポンチ絵スライドになってしまうことを回避する画期的な機能もあります
}
\details{
XeLaTeX + beamer + XeCJK
}
